requirements.txt
anthropic>=0.40.0
google-genai>=0.8.0
openai>=1.0.0
Pillow>=10.0.0
python-dotenv>=1.0.0
